init clarity restructure, not built/tested
